Implement host deletion, cleanup database.
- Removed 'deleted_systems' table is not needed, since we are querying
inventory for systems, and can detect out-of-order deletions that way.

This table was used as a guard against accidentally reordered upload,
and delete messages. If in a rare case, a delete message was processed
before an upload message for same host, it would mean that host would
be deleted and re-created. We can detect this when processing an upload
message. If the call to get 'system_profile' returns 404, host is deleted.

- Re-organized the handlers, to make event parsing code shared.